requirements.txt
Flask==2.3.2
pandas==2.2.0
SQLAlchemy==2.0.19
python-dateutil==2.8.2
pytest==7.4.0